php stan and php unit test added

+++ uninstall.php
@@ -1,30 +1,34 @@
 <?php
-	/**
-	 * Uninstall Plugin
-	 *
-	 * Deletes all plugin settings.
-	 *
-	 * @package    StorePress/Base
-	 * @since      1.0.0
-	 */
+/**
+ * Uninstall Plugin
+ *
+ * Deletes all plugin settings.
+ *
+ * @package    StorePress/Base
+ * @since      1.0.0
+ */
 
-	namespace StorePress\Base;
+declare( strict_types=1 );
 
-	/**
-	 * This file describes all uninstall logic.
-	 */
+namespace StorePress\Base;
+
+/**
+ * This file describes all uninstall logic.
+ */
 if ( ! defined( 'ABSPATH' ) ) {
 	exit;
 }
 
-	// Exit if accessed directly.
+// Exit if accessed directly.
 if ( ! defined( 'WP_UNINSTALL_PLUGIN' ) ) {
 	exit;
 }
 
-	$storepress_settings = get_option( 'plugin_settings' );
+$storepress_settings = get_option( 'plugin_settings' );
 
-	// Delete all settings.
-if ( $storepress_settings['plugin_settings']['remove_on_uninstall'] ) {
+// Delete all settings.
+if ( is_array( $storepress_settings )
+	&& isset( $storepress_settings['plugin_settings']['remove_on_uninstall'] )
+) {
 	delete_option( 'plugin_settings' );
 }
